Fully removed the helper
